pyproject.toml
[tool.poetry]
name = "robometrics"
version = "0.1.0"
description = "A unified framework for metrics evaluation for robotic manipulation"
authors = ["Adam Fishman"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.15"
pyyaml = ">5.3" # there is a bug in installing pyyaml < 6.0 in python 3.10
geometrout = { version = "^0.1.0.10", optional = true }
urchin = { version = "^0.0.26", optional = true }

[tool.poetry.extras]
evaluator = ["geometrout", "urchin"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
